HTTP module bug fix: 'query' part was missed in the http request first line
http: request first line should contain 'path' and 'query' parts
log: avoid streamsize to size_t warning

+++ include/hive/log.hpp
@@ -624,7 +624,7 @@ public:
     */
     size_t getMaxFileSize() const
     {
-        return m_maxFileSize;
+        return size_t(m_maxFileSize);
     }
 
 public:
